+++ packages/provider/test/provider.test.js
@@ -181,6 +181,21 @@ describe('Provider', function () {
   });
 
   describe('#getItem', function () {
+    const ancestorFallback = () => <div>I feel so old</div>;
+    const customFallback = () => <div>¯\_(ツ)_/¯</div>;
+
+    function useCustomFallbackProvider(opts) {
+      const fallback = opts ? opts.fallback : customFallback
+      wrapper = shallow(
+        <Provider uri='http://example.com/500' parser={ parser } fallback={ fallback }>
+          <div>Example</div>
+        </Provider>,
+        opts && opts.options
+      );
+  
+      provider = wrapper.instance();
+    }
+
     it('queues the action if we are currently fetching a resource', function () {
       assume(provider.queue).is.length(0);
 
@@ -231,6 +246,54 @@ describe('Provider', function () {
       });
     });
 
+    it('returns the custom fallback and error in case of error state', function (next) {
+      useCustomFallbackProvider({ fallback: customFallback });
+      provider.state.readyState = READYSTATES.LOADED;
+      provider.state.error = new Error('Something went wrong');
+
+      provider.getItem('name', (err, data) => {
+        assume(err).is.a('error');
+        assume(err.message).equals('Something went wrong');
+
+        assume(data).equals(customFallback);
+
+        next();
+      });
+    });
+
+    it('returns the local custom fallback and error in case of error state', function (next) {
+      useCustomFallbackProvider({
+        fallback: customFallback,
+        options: { context: { Fallback: ancestorFallback }}
+      });
+      provider.state.readyState = READYSTATES.LOADED;
+      provider.state.error = new Error('Something went wrong');
+
+      provider.getItem('name', (err, data) => {
+        assume(err).is.a('error');
+        assume(err.message).equals('Something went wrong');
+
+        assume(data).equals(customFallback);
+
+        next();
+      });
+    });
+
+    it('returns the ancestor custom fallback and error in case of error state', function (next) {
+      useCustomFallbackProvider({ options: { context: { Fallback: ancestorFallback }}});
+      provider.state.readyState = READYSTATES.LOADED;
+      provider.state.error = new Error('Something went wrong');
+
+      provider.getItem('name', (err, data) => {
+        assume(err).is.a('error');
+        assume(err.message).equals('Something went wrong');
+
+        assume(data).equals(ancestorFallback);
+
+        next();
+      });
+    });
+
     it('returns the fallback and error when receiving an unknown name', function (next) {
       provider.state.readyState = READYSTATES.LOADED;
 
@@ -239,6 +302,52 @@ describe('Provider', function () {
         assume(err.message).equals('Unknown SVG requested');
 
         assume(data).equals(Fallback);
+
+        next();
+      });
+    });
+
+    it('returns the custom fallback and error when receiving an unknown name', function (next) {
+      useCustomFallbackProvider({ fallback: customFallback });
+      provider.state.readyState = READYSTATES.LOADED;
+
+      provider.getItem('name', (err, data) => {
+        assume(err).is.a('error');
+        assume(err.message).equals('Unknown SVG requested');
+
+        assume(data).equals(customFallback);
+
+        next();
+      });
+    });
+
+    it('returns the local custom fallback and error when receiving an unknown name', function (next) {
+      useCustomFallbackProvider({
+        fallback: customFallback,
+        options: { context: { Fallback: ancestorFallback }}
+      });
+      provider.state.readyState = READYSTATES.LOADED;
+
+      provider.getItem('name', (err, data) => {
+        assume(err).is.a('error');
+        assume(err.message).equals('Unknown SVG requested');
+
+        assume(data).equals(customFallback);
+
+        next();
+      });
+    });
+
+
+    it('returns the ancestor custom fallback when no local fallback and error when receiving an unknown name', function (next) {
+      useCustomFallbackProvider({ options: { context: { Fallback: ancestorFallback }}});
+      provider.state.readyState = READYSTATES.LOADED;
+
+      provider.getItem('name', (err, data) => {
+        assume(err).is.a('error');
+        assume(err.message).equals('Unknown SVG requested');
+
+        assume(data).equals(ancestorFallback);
 
         next();
       });